Moves AccountsIndexIterator to submodule
